requirements.txt
pandas>=2.0.0
python-dotenv>=1.0.0
requests>=2.28.0
google-generativeai>=0.3.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0
numpy>=1.24.0
ijson>=3.2.0
chardet>=5.0.0
dealcloud-sdk>=1.0.0
